schonsteil/mysite/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.14.0
django-pictures==1.0.1
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
geographiclib==2.0
geopy==2.3.0
gpx-cmd-tools==0.0.1
gpxpy==1.5.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==1.7.1
pyproj==3.5.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
shapely==2.0.1
sorl-thumbnail==12.9.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
